Tic tac toe game (C programming language)
Game consists multiple choices such as 3x3 and 4x4 board, playing with AI bots (easy, medium and hard level) or playing with another human, playing tournament or playing the single game.

The program remembers users' scores and creates a file with their information and scores. The user has the chance to look at their files after the games are finished.